+++ lib/iris/tests/unit/fileformats/netcdf/test_save.py
@@ -143,7 +143,7 @@ class Test_fill_value(tests.IrisTest):
         # Test that when no fill_value argument is passed, the fill_value
         # argument to Saver.write is None or not present.
         cubes = self._make_cubes()
-        with mock.patch("iris.fileformats.netcdf.Saver") as Saver:
+        with mock.patch("iris.fileformats.netcdf.saver.Saver") as Saver:
             save(cubes, "dummy.nc")
 
         # Get the Saver.write mock
@@ -161,7 +161,7 @@ class Test_fill_value(tests.IrisTest):
         # that value is passed to each call to Saver.write
         cubes = self._make_cubes()
         fill_value = 12345.0
-        with mock.patch("iris.fileformats.netcdf.Saver") as Saver:
+        with mock.patch("iris.fileformats.netcdf.saver.Saver") as Saver:
             save(cubes, "dummy.nc", fill_value=fill_value)
 
         # Get the Saver.write mock
@@ -178,7 +178,7 @@ class Test_fill_value(tests.IrisTest):
         # each element is passed to separate calls to Saver.write
         cubes = self._make_cubes()
         fill_values = [123.0, 456.0, 789.0]
-        with mock.patch("iris.fileformats.netcdf.Saver") as Saver:
+        with mock.patch("iris.fileformats.netcdf.saver.Saver") as Saver:
             save(cubes, "dummy.nc", fill_value=fill_values)
 
         # Get the Saver.write mock
@@ -195,7 +195,7 @@ class Test_fill_value(tests.IrisTest):
         # that value is passed to calls to Saver.write
         cube = Cube(["abc", "def", "hij"])
         fill_value = "xyz"
-        with mock.patch("iris.fileformats.netcdf.Saver") as Saver:
+        with mock.patch("iris.fileformats.netcdf.saver.Saver") as Saver:
             save(cube, "dummy.nc", fill_value=fill_value)
 
         # Get the Saver.write mock
@@ -211,7 +211,7 @@ class Test_fill_value(tests.IrisTest):
         # is passed as the fill_value argument, an error is raised
         cubes = self._make_cubes()
         fill_values = [1.0, 2.0, 3.0, 4.0]
-        with mock.patch("iris.fileformats.netcdf.Saver"):
+        with mock.patch("iris.fileformats.netcdf.saver.Saver"):
             with self.assertRaises(ValueError):
                 save(cubes, "dummy.nc", fill_value=fill_values)
